+++ tests/Unit/Domain/SignUp/UseCases/ResendActivationEmailTest.php
@@ -2,70 +2,60 @@
 
 namespace Tests\Unit\Domain\SignUp\UseCases;
 
-use Domain\SignUp\UseCases\CreateUnactivatedAccount;
+use Domain\SignUp\Exceptions\AccountAlreadyActivatedException;
+use Domain\SignUp\UseCases\ResendActivationEmail;
 use Entities\Models\Eloquent\Account;
 use Entities\Notifications\AccountActivationNotification;
 use Illuminate\Support\Facades\Notification;
 use Library\SignedURL\Adapters\StubSignedURLGenerator;
 use Library\SignedURL\SignedURLGenerator;
 use Repositories\AccountRepository;
-use Shared\Groups\GroupsManager;
 use Tests\TestCase;
 
-class CreateUnactivatedAccountUseCaseTest extends TestCase
+class ResendActivationEmailTest extends TestCase
 {
     private AccountRepository $accountRepository;
-    private GroupsManager $groupsManager;
     private SignedURLGenerator $signedURLGenerator;
-    private CreateUnactivatedAccount $useCase;
+    private ResendActivationEmail $useCase;
 
     protected function setUp(): void
     {
         parent::setUp();
 
         $this->accountRepository = \Mockery::mock(AccountRepository::class);
-        $this->groupsManager = \Mockery::mock(GroupsManager::class);
         $this->signedURLGenerator = new StubSignedURLGenerator(outputURL: 'url');
 
-        $this->useCase = new CreateUnactivatedAccount(
+        $this->useCase = new ResendActivationEmail(
             accountRepository: $this->accountRepository,
-            groupsManager: $this->groupsManager,
             signedURLGenerator: $this->signedURLGenerator,
         );
 
         Notification::fake();
     }
 
-    public function test_creates_unactivated_account()
+    public function test_throws_exception_if_account_already_activated()
     {
-        $email = 'email';
-        $username = 'username';
-        $password = 'password';
-        $ip = 'ip';
+        $account = Account::factory()->create(['activated' => true]);
 
         $this->accountRepository
-            ->shouldReceive('create')
-            ->with($email, $username, $password, $ip)
-            ->andReturn(new Account());
+            ->shouldReceive('getByEmail')
+            ->with($account->email)
+            ->andReturn($account);
 
-        $this->groupsManager
-            ->shouldReceive('addToDefaultGroup');
+        $this->expectException(AccountAlreadyActivatedException::class);
 
-        $this->useCase->execute($email, $username, $password, $ip);
+        $this->useCase->execute(email: $account->email);
     }
 
     public function test_sends_notification()
     {
-        $account = Account::factory()->create();
+        $account = Account::factory()->create(['activated' => false]);
 
         $this->accountRepository
-            ->shouldReceive('create')
+            ->shouldReceive('getByEmail')
             ->andReturn($account);
 
-        $this->groupsManager
-            ->shouldReceive('addToDefaultGroup');
-
-        $this->useCase->execute('email', 'username', 'password', 'ip');
+        $this->useCase->execute('email');
 
         Notification::assertSentTo($account, AccountActivationNotification::class);
     }
